reafactor: move bench/lsp harness to test_util
