+++ test_util/Cargo.toml
@@ -12,6 +12,7 @@ name = "test_server"
 path = "src/test_server.rs"
 
 [dependencies]
+anyhow = "1.0.40"
 async-stream = "0.3.0"
 bytes = "1.0.1"
 futures = "0.3.13"
@@ -20,6 +21,7 @@ lazy_static = "1.4.0"
 os_pipe = "0.9.2"
 regex = "1.4.3"
 serde = { version = "1.0.125", features = ["derive"] }
+serde_json = "1.0.64"
 tempfile = "3.2.0"
 tokio = { version = "1.6.0", features = ["full"] }
 tokio-rustls = "0.22.0"
